// src/app/model/coronaVirus.ts
export class CoronaVirus {
    id: string;
    continent: string;
    country: string;
    provinceState: string;
    population: number;
    lastUpdateDisplay: string;
    lastUpdate: number;
    lastChangeDisplay: string;
    lastChange: number;
    lat: number;
    long_: number;
    confirmed: number;
    recovered: number;
    deaths: number;
    active: number;
    critical: number;
}